requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
dnspython==1.15.0
idna==2.6
pyzmq==16.0.3
requests==2.18.4
requests-file==1.4.2
six==1.11.0
tldextract==2.2.0
urllib3==1.22
